fix criterion debug page crash issue
